implement login and registration pages; add opportunity component to feed and publish pages; refactor layout and styling across various components
