+++ frontend/src/AppPage.jsx
@@ -1,5 +1,5 @@
 import {BrowserRouter, Route, Routes} from 'react-router-dom';
-import HomePage from './pages/HomePage';
+import LoginPage from './pages/LoginPage';
 import FeedPage from './pages/FeedPage';
 import ProfilePage from './pages/ProfilePage';
 import PublishPage from './pages/PublishPage';
@@ -13,7 +13,7 @@ const App = () => {
 		<BrowserRouter>
 			<Header />
 			<Routes>
-				<Route path="/" element={<HomePage />} />
+				<Route path="/login" element={<LoginPage />} />
 				<Route path="/feed" element={<FeedPage />} />
 				<Route path="/profile" element={<ProfilePage />} />
 				<Route path="/publish" element={<PublishPage />} />
