+++ frontend/src/index.css
@@ -9,6 +9,7 @@
     --color-textbody: #333333;
     --color-info: #666666;
     --color-divider: #e6e6e6;
+    --color-whitered: #FDEBED;  
 }
 
 html, body {
